#BAH-921 | Refactor Bahmni IE Apps to migrate API-related classes to the api folder
