+++ api/pom.xml
@@ -14,17 +14,6 @@
 	<description>API project for bahmni.ie.apps</description>
 
 	<dependencies>
-	
-		<!-- 
-		Add other dependencies from parent's pom:
-		<dependency>
-			<groupId>org.other.library</groupId> 
-			<artifactId>library-name</artifactId>
-		</dependency> 
-		-->
-
-		
-		
 
 		<!-- Begin OpenMRS core -->
 		
@@ -59,6 +48,12 @@
 			<artifactId>openmrs-test</artifactId>
 			<type>pom</type>
 			<scope>test</scope>
+		</dependency>
+
+		<dependency>
+			<groupId>org.json</groupId>
+			<artifactId>json</artifactId>
+			<version>${jsonVersion}</version>
 		</dependency>
 				
 		<!-- End OpenMRS core -->
