+++ api/src/main/java/org/openmrs/module/bahmni/ie/apps/model/BahmniForm.java
@@ -1,0 +1,44 @@
+package org.openmrs.module.bahmni.ie.apps.model;
+
+public class BahmniForm {
+
+	private String name;
+
+	private String uuid;
+
+	private String version;
+
+	private boolean published;
+
+	public String getName() {
+		return name;
+	}
+
+	public void setName(String name) {
+		this.name = name;
+	}
+
+	public String getUuid() {
+		return uuid;
+	}
+
+	public void setUuid(String uuid) {
+		this.uuid = uuid;
+	}
+
+	public String getVersion() {
+		return version;
+	}
+
+	public void setVersion(String version) {
+		this.version = version;
+	}
+
+	public boolean isPublished() {
+		return published;
+	}
+
+	public void setPublished(boolean published) {
+		this.published = published;
+	}
+}
